# docs/swagger.yaml
basePath: /v1
definitions:
  models.CaptchaResponse:
    properties:
      captchaId:
        description: 验证码Id
        type: string
      imageUrl:
        description: 验证码图片url
        type: string
    type: object
  models.Goods:
    properties:
      goodsId:
        description: 商品id
        type: integer
      goodsName:
        description: 商品名字
        type: string
      price:
        description: 商品价格
        type: string
      stock:
        description: 商品库存
        type: integer
      subject:
        description: 商品描述
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        description: 创建时间戳
        type: integer
      email:
        description: 用户邮箱
        type: string
      first_name:
        description: 用户名
        type: string
      gender:
        description: 性别
        type: string
      id:
        description: 用户id
        type: integer
      last_name:
        description: 用户名
        type: string
      mobile_phone:
        description: 移动电话
        type: string
      password:
        description: 用户密码
        type: string
      tenant_id:
        description: tenant_id
        type: string
      updated_at:
        description: 创建时间戳
        type: integer
      user_name:
        description: 用户名
        type: string
      uuid:
        description: 用户uuid
        type: string
    required:
    - email
    - mobile_phone
    - password
    - tenant_id
    - user_name
    type: object
...
  swag.ResultContChangePass:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
...
  swag.ResultContGetCaptchaImage:
    properties:
      data:
        $ref: '#/definitions/models.CaptchaResponse'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
...
  swag.ResultContGoods:
    properties:
      data:
        $ref: '#/definitions/models.Goods'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContGoodsList:
    properties:
      data:
        description: 商品列表
        items:
          $ref: '#/definitions/models.Goods'
        type: array
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContSingIn:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContSingUp:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContTest:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
...
  swag.ResultContVerifyCaptcha:
    properties:
      data:
        $ref: '#/definitions/models.CaptchaResponse'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
host: localhost:9090
info:
  contact: {}
  description: XXXX项目描述
  title: XXXX项目名称-Swagger Example API
  version: "1.0"
paths:
...
  /captcha/{captchaId}:
    get:
      description: 获取验证码 GetCaptcha
      parameters:
      - description: 验证码Id
        in: path
        name: captchaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGetCaptchaImage'
      summary: 获取验证码图片
  /captcha/verify/{captchaId}/{value}:
    get:
      description: 获取验证码 GetCaptcha
      parameters:
      - description: 验证码Id
        in: path
        name: captchaId
        required: true
        type: integer
      - description: 验证码value
        in: path
        name: value
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContVerifyCaptcha'
      summary: 验证 验证码码图
  /changepass:
    put:
      description: 输入用户名及密登录
      parameters:
      - description: 用户邮箱
        in: path
        name: email
        required: true
        type: string
      - description: 旧密码
        in: path
        name: oldpassword
        required: true
        type: string
      - description: 新密码
        in: path
        name: newpassword
        required: true
        type: string
      - description: 确认新密码
        in: path
        name: confirmpassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContChangePass'
      summary: 修改密码
  /index/goodslist:
    get:
      description: 显示商品列表
      parameters:
      - description: 分类id
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGoodsList'
      summary: 商品列表
  /index/goodsone/{goodsid}:
    get:
      description: 显示商品的详情
      parameters:
      - description: 商品id
        in: path
        name: goodsid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGoods'
      summary: 商品详情
  /login:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingIn'
      summary: 系统超级管理登录
  /singin:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 承租
        in: path
        name: tenant
        required: true
        type: string
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingIn'
      summary: 系统登录
  /singup:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 承租户tenant_id
        in: path
        name: tenant_id
        required: true
        type: integer
      - description: 用户邮箱
        in: path
        name: email
        required: true
        type: string
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingUp'
      summary: 注册用户
...
  /test:
    get:
      description: Test===
      parameters:
      - description: check_in
        in: path
        name: check_in
        required: true
        type: string
      - description: check_out
        in: path
        name: check_out
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContTest'
      summary: Test
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"